requirements.txt
mcp[cli]>=1.0.0
google-genai>=1.0.0
pydantic>=2.0.0
httpx>=0.25.0
